requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=-
#
noise==1.2.2
    # via -r requirements.in
numpy==1.21.3
    # via -r requirements.in
perlin-noise==1.8
    # via -r requirements.in
pillow==8.4.0
    # via -r requirements.in
